+++ apps/test/__init__.py
@@ -1,12 +1,13 @@
 import importlib, inspect
-from system import utils
+from system import utils, homepage
+
+INACCESSIBLE = True
 
 
 def launch():
     app_name = input("Enter the name of the app you want to test.\nEnter nothing to exit.\n>")
     if app_name == "":
-        from system import homepage
-        homepage.launch()
+        return homepage.launch()
     try:
         app = importlib.import_module(f"testing.{app_name}")
         if not inspect.isfunction(app.launch):
@@ -16,13 +17,10 @@ def launch():
         return launch()
     except AttributeError:
         print("App has no `launch` function!")
-        from system import homepage
         return homepage.launch()
     except Exception as err:
         if Exception.__str__() == "WRONG_TYPE":
             print("`launch` must be a function!")
-            from system import homepage
-            homepage.launch()
+            return homepage.launch()
         print(err)
-        from system import homepage
         return homepage.launch()