+++ installation/__init__.py
@@ -1,4 +1,5 @@
-import shutil, os
+import shutil, os, importlib, sys
+sys.path.append("../")
 
 def install(app_name, path="apps"):
     apps = os.listdir("installation")
@@ -18,7 +19,7 @@ def install(app_name, path="apps"):
         if app_name == a:
             shutil.copytree(os.path.join("installation", app_name), os.path.join(path, app_name))
             break
-    
+
     else:
         raise ModuleNotFoundError("App not found or installed.")
 
@@ -27,13 +28,20 @@ def uninstall(app_name, path):
     apps = os.listdir(path)
     a = 0
     while a < len(apps):
-        if apps[a] in ("__init__.py", "__pycache__", "user", "test"):
+        if apps[a] in ("__init__.py", "__pycache__", "user"):
             del apps[a]
         a += 1
     del a
     for a in apps:
         if app_name == a:
-            shutil.rmtree(os.path.join(path, app_name))
+            try:
+                module = importlib.import_module(f".{app_name}", path)
+                if module.INACCESSIBLE:
+                    raise ImportError("App inaccessible.")
+            except AttributeError:
+                pass
+            # shutil.rmtree(os.path.join(path, app_name))
+            print(os.path.join(path, app_name))
             break
 
     else:
@@ -57,7 +65,7 @@ def search(app_name, user_path):
     for a in apps:
         if app_name == a:
             return "public"
-    
+
     for a in user_apps:
         if app_name == a:
             return "local"