+++ main.py
@@ -1,9 +1,8 @@
-from system import account
+from system import account, utils
 
-print("""
-+------------- Rune OS 3.7.1 -------------+
-|      Made by @IY314 and @mrfoogles      |
-|    https://github.com/IY314/Rune-OS     |
-+-----------------------------------------+
-""")
+VERSION = "Rune OS 3.7.2"
+CREDITS = "Made by @IY314 and @mrfoogles"
+LINK = "https://github.com/IY314/Rune-OS"
+
+print(utils.make_box(VERSION, CREDITS, LINK))
 account.login(False)